Update gulpfile.js
Functionality is the same but fixed the overall base of the program. Using for example an object with files. Also using some plugins for images that fixes all updates of adding or removing images. Fixed a default task that will run htmlTask, jsTask, scssTask and imgTask once and after that watchTask. Just run it with command gulp.